# public/openapi.yaml
openapi: "3.0.1"
info:
  title: "Factorio Item Browser API"
  description:
    This API provides access to the mod data of the Factorio Item Browser, including item details, recipe details as
    well as any translation available from the mods.
  contact:
    name: "BluePsyduck"
    email: "bluepsyduck@gmx.com"
  version: "1.1.0"
servers:
  - url: "https://www.factorio-item-browser.com/api"
    description: "Live API server"
...
paths:
...
  /item/ingredient:
    post:
      summary:
        Provides all recipes in which a specific item as an ingredient of.
      description:
        This request will return all recipes with a specific item as one of its ingredients. The recipes in the response
        will be grouped together by their name, so that normal and expensive versions will always be returned together,
        if both are available. The parameters "numberOfResults" and "indexOfFirstResult" are applied to these groups,
        not on the recipes itself.
      operationId: "item.ingredient"
      tags:
        - Items
      parameters:
        - $ref: "openapi/parameters.yaml#/Authorization"
        - $ref: "openapi/parameters.yaml#/Accept-Language"
      requestBody:
        $ref: "openapi/requestBodies.yaml#/itemRecipes"
      responses:
        "200":
          $ref: "openapi/responses.yaml#/itemIngredient"
        "400":
          $ref: "openapi/errorResponses.yaml#/badRequest"
        "401":
          $ref: "openapi/errorResponses.yaml#/unauthorized"
        "404":
          $ref: "openapi/errorResponses.yaml#/itemNotFound"
        "500":
          $ref: "openapi/errorResponses.yaml#/unexpectedError"
      security:
        - jwt: []
  /item/product:
    post:
      summary:
        Provides all recipes which produce a specific item.
      description:
        This request will return all recipes with a specific item as one of its products. The recipes in the response
        will be grouped together by their name, so that normal and expensive versions will always be returned together,
        if both are available. The parameters "numberOfResults" and "indexOfFirstResult" are applied to these groups,
        not on the recipes itself.
      operationId: "item.product"
      tags:
        - Items
      parameters:
        - $ref: "openapi/parameters.yaml#/Authorization"
        - $ref: "openapi/parameters.yaml#/Accept-Language"
      requestBody:
        $ref: "openapi/requestBodies.yaml#/itemRecipes"
      responses:
        "200":
          $ref: "openapi/responses.yaml#/itemProduct"
        "400":
          $ref: "openapi/errorResponses.yaml#/badRequest"
        "401":
          $ref: "openapi/errorResponses.yaml#/unauthorized"
        "404":
          $ref: "openapi/errorResponses.yaml#/itemNotFound"
        "500":
          $ref: "openapi/errorResponses.yaml#/unexpectedError"
      security:
        - jwt: []
...
components:
...
  securitySchemes:
    jwt:
      description: "JWT Token for the Factorio Item Browser API. Use the /auth request to receive a token."
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"